Move env settings to gzserver action
